JupyterLab instance based on AiiDAlab Docker stack
A Dockerfile based on the AiiDAlab image 'base-with-services'.
On top of this a new virtual environment is created, the ipykernel is
installed as well as ExecFlow.
Then a new ipykernel is created called "OpenModel Execflow".

Temporarily the ExecFlow version used is installed from the GitHub
repository, and the branch 'cwa/fix-26...' is used. This specific branch
usage should be removed.

Co-authored-by: Thomas Hagelien <thomas.f.hagelien@sintef.no>